move(board): dto 패키지 내부로 파일 이동
