+++ src/main/java/com/example/workoutmate/domain/board/service/BoardService.java
@@ -1,7 +1,7 @@
 package com.example.workoutmate.domain.board.service;
 
-import com.example.workoutmate.domain.board.controller.dto.BoardRequestDto;
-import com.example.workoutmate.domain.board.controller.dto.BoardResponseDto;
+import com.example.workoutmate.domain.board.dto.BoardRequestDto;
+import com.example.workoutmate.domain.board.dto.BoardResponseDto;
 import com.example.workoutmate.domain.board.entity.Board;
 import com.example.workoutmate.domain.board.entity.SportType;
 import com.example.workoutmate.domain.board.repository.BoardRepository;
